+++ core/inc/com/centreon/broker/multiplexing/engine.hh
@@ -19,12 +19,12 @@
 #ifndef CCB_MULTIPLEXING_ENGINE_HH
 #define CCB_MULTIPLEXING_ENGINE_HH
 
+#include <deque>
 #include <list>
 #include <memory>
 #include <mutex>
-#include <queue>
+#include <asio.hpp>
 
-#include "com/centreon/broker/multiplexing/hooker.hh"
 #include "com/centreon/broker/namespace.hh"
 #include "com/centreon/broker/persistent_cache.hh"
 
@@ -48,17 +48,16 @@ class muxer;
  *  The instance initialization/deinitialization are guarded by a mutex
  *  _load_m. It is only used for that purpose.
  *
- *  This class is the root of events dispatching. Events arrive from a stream
+ *  This class is the root of events dispatching. Events arrive from a stream,
  *  are transfered to a muxer and then to engine (at the root of the tree).
- *  This one then sends the event to all its children. Each muxer receives
- *  the event and sends it to its stream.
+ *  This one then sends events to all its children. Each muxer receives
+ *  these events and sends them to its stream.
  *
  *  The engine has three states:
- *  * switched off, the 'write' function points to a _nop() function. All event
- *    that could be received is lost by the engine. This state is possible only
- *    when the engine is started or during tests.
- *  * running, the 'write' function points to a _write() function that sends
- *    received events to all the muxers beside.
+ *  * not started. All event that could be received is lost by the engine.
+ *    This state is possible only when the engine is started or during tests.
+ *  * running, received events are dispatched to all the muxers beside. This
+ *    is done asynchronously.
  *  * stopped, the 'write' function points to a _write_to_cache_file() funtion.
  *    When broker is stopped, before it to be totally stopped, events are
  *    written to a cache file ...unprocessed... This file will be re-read at the
@@ -67,33 +66,32 @@ class muxer;
  *  @see muxer
  */
 class engine {
+  static std::mutex _load_m;
   static engine* _instance;
+
+  enum state { not_started, running, stopped };
+  state _state;
+  asio::io_context::strand _strand;
+
   std::unique_ptr<persistent_cache> _cache_file;
 
-  // Data queue.
-  std::queue<std::shared_ptr<io::data>> _kiew;
-
-  // Hooks
-  std::vector<std::pair<hooker*, bool>> _hooks;
-  std::vector<std::pair<hooker*, bool>>::iterator _hooks_begin;
-  std::vector<std::pair<hooker*, bool>>::iterator _hooks_end;
-
-  // Mutex to lock _kiew and _hooks
+  // Mutex to lock _kiew and _state
   std::mutex _engine_m;
+
+  // Data queue.
+  std::deque<std::shared_ptr<io::data>> _kiew;
 
   // Subscriber.
   std::vector<muxer*> _muxers;
-  std::mutex _muxers_m;
 
-  static std::mutex _load_m;
+  // Statistics.
+  uint32_t _unprocessed_events;
+
+  std::atomic_bool _sending_to_subscribers;
 
   engine();
   std::string _cache_file_path() const;
-  void _nop(std::shared_ptr<io::data> const& d);
   void _send_to_subscribers();
-  void _write(std::shared_ptr<io::data> const& d);
-  void _write_to_cache_file(std::shared_ptr<io::data> const& d);
-  void _publish(std::shared_ptr<io::data> const& d);
 
   void (engine::*_write_func)(std::shared_ptr<io::data> const&);
 
@@ -102,16 +100,15 @@ class engine {
   static void unload();
   static engine& instance();
 
-  engine(engine const&) = delete;
-  engine& operator=(engine const&) = delete;
+  engine(const engine&) = delete;
+  engine& operator=(const engine&) = delete;
   ~engine() noexcept = default;
+
   void clear();
   void publish(const std::shared_ptr<io::data>& d);
   void publish(const std::list<std::shared_ptr<io::data>>& to_publish);
   void start();
   void stop();
-  void hook(hooker& h, bool with_data = true);
-  void unhook(hooker& h);
   void subscribe(muxer* subscriber);
   void unsubscribe(muxer* subscriber);
 };