+++ core/test/multiplexing/subscriber/ctor_default.cc
@@ -20,6 +20,7 @@
 #include <gtest/gtest.h>
 
 #include <cstring>
+#include <thread>
 
 #include "com/centreon/broker/config/applier/init.hh"
 #include "com/centreon/broker/io/events.hh"
@@ -64,8 +65,15 @@ TEST_F(SubscriberTest, DefaultConstructor) {
   data->append(MSG);
   s.get_muxer().write(std::static_pointer_cast<io::data>(data));
 
-  // Fetch event.
-  s.get_muxer().read(event, 0);
+  bool ret;
+  int count = 0;
+  do {
+    // Fetch event.
+    ret = s.get_muxer().read(event, 0);
+    std::this_thread::sleep_for(std::chrono::milliseconds(10));
+    count++;
+  } while (!ret && count < 100);
+
   ASSERT_TRUE(event);
   ASSERT_EQ(event->type(), io::raw::static_type());
   ASSERT_TRUE(strncmp(std::static_pointer_cast<io::raw>(event)->const_data(),